+++ eventlog-live-otelcol/src/GHC/Eventlog/Live/Otelcol/Config/Types.hs
@@ -10,6 +10,8 @@ module GHC.Eventlog.Live.Otelcol.Config.Types (
   Metrics (..),
   Spans (..),
   AggregationStrategy (..),
+  ExportStrategy (..),
+  toExportBatches,
   HeapAllocatedMetric (..),
   BlocksSizeMetric (..),
   HeapSizeMetric (..),
@@ -23,10 +25,17 @@ module GHC.Eventlog.Live.Otelcol.Config.Types (
   ThreadStateSpan (..),
 ) where
 
+import Data.Aeson.Encoding qualified as AE
 import Data.Aeson.Types (Encoding, FromJSON (..), Options (..), Parser, SumEncoding (..), ToJSON (..), Value (..), camelTo2, defaultOptions, genericParseJSON, genericToEncoding, genericToJSON)
+import Data.Aeson.Types qualified as AT
+import Data.Char (isDigit)
 import Data.Text (Text)
+import Data.Text qualified as T
 import GHC.Generics (Generic)
+import GHC.Records (HasField (..))
 import Language.Haskell.TH.Lift.Compat (Lift)
+import Text.ParserCombinators.ReadP (ReadP)
+import Text.ParserCombinators.ReadP qualified as P
 
 {- |
 The configuration for @eventlog-live-otelcol@.
@@ -111,33 +120,119 @@ instance ToJSON Spans where
   toEncoding = genericToEncoding encodingOptions
 
 {- |
-The options for metric aggregation.
+The options for metric aggregation strategies.
 -}
 data AggregationStrategy
-  = AggregationStrategyByBatch
+  = AggregationStrategyBool {isOn :: !Bool}
   | AggregationStrategyByBatches {byBatches :: !Int}
   deriving (Generic, Lift)
 
+{- |
+Internal helper.
+A `ReadP` style parser for `AggregationStrategy`.
+-}
+readPAggregationStrategy :: ReadP AggregationStrategy
+readPAggregationStrategy =
+  AggregationStrategyByBatches . read <$> P.munch1 isDigit <* P.char 'x'
+
 instance FromJSON AggregationStrategy where
   parseJSON :: Value -> Parser AggregationStrategy
-  parseJSON = genericParseJSON aggregationStrategyEncodingOptions
+  parseJSON = \case
+    AT.Bool isOn ->
+      pure $ AggregationStrategyBool isOn
+    AT.String txt
+      | [(aggregationStrategy, "")] <-
+          P.readP_to_S readPAggregationStrategy (T.unpack txt) ->
+          pure aggregationStrategy
+    value -> AT.unexpected value
 
 instance ToJSON AggregationStrategy where
   toJSON :: AggregationStrategy -> Value
-  toJSON = genericToJSON aggregationStrategyEncodingOptions
+  toJSON = \case
+    AggregationStrategyBool{..} ->
+      Bool isOn
+    AggregationStrategyByBatches{..} ->
+      String (T.pack (show byBatches <> "x"))
   toEncoding :: AggregationStrategy -> Encoding
-  toEncoding = genericToEncoding aggregationStrategyEncodingOptions
+  toEncoding = \case
+    AggregationStrategyBool{..} ->
+      AE.bool isOn
+    AggregationStrategyByBatches{..} ->
+      AE.text (T.pack (show byBatches <> "x"))
+
+{- |
+The options for export strategies.
+-}
+data ExportStrategy
+  = ExportStrategyBool {isOn :: Bool}
+  | ExportStrategyByBatches {byBatches :: Int}
+  deriving (Generic, Lift)
+
+{- |
+Check whether or not a processor is enabled based on its export strategy.
+-}
+isEnabled :: Maybe ExportStrategy -> Bool
+isEnabled = \case
+  Nothing -> False
+  Just ExportStrategyBool{..} -> isOn
+  Just ExportStrategyByBatches{..} -> byBatches >= 1
+
+{- |
+Convert an `ExportStrategy` to a number of batches.
+-}
+toExportBatches :: Maybe ExportStrategy -> Int
+toExportBatches = \case
+  Nothing -> 0
+  Just ExportStrategyBool{..} -> if isOn then 1 else 0
+  Just ExportStrategyByBatches{..} -> byBatches
+
+{- |
+Internal helper.
+A `ReadP` style parser for `ExportStrategy`.
+-}
+readPExportStrategy :: ReadP ExportStrategy
+readPExportStrategy =
+  ExportStrategyByBatches . read <$> P.munch1 isDigit <* P.char 'x'
+
+instance FromJSON ExportStrategy where
+  parseJSON :: Value -> Parser ExportStrategy
+  parseJSON = \case
+    AT.Bool isOn ->
+      pure $ ExportStrategyBool isOn
+    AT.String txt
+      | [(aggregationStrategy, "")] <-
+          P.readP_to_S readPExportStrategy (T.unpack txt) ->
+          pure aggregationStrategy
+    value -> AT.unexpected value
+
+instance ToJSON ExportStrategy where
+  toJSON :: ExportStrategy -> Value
+  toJSON = \case
+    ExportStrategyBool{..} ->
+      Bool isOn
+    ExportStrategyByBatches{..} ->
+      String (T.pack (show byBatches <> "x"))
+  toEncoding :: ExportStrategy -> Encoding
+  toEncoding = \case
+    ExportStrategyBool{..} ->
+      AE.bool isOn
+    ExportStrategyByBatches{..} ->
+      AE.text (T.pack (show byBatches <> "x"))
 
 {- |
 The configuration options for `GHC.Eventlog.Live.Machine.Analysis.Heap.processHeapAllocatedData`.
 -}
 data HeapAllocatedMetric = HeapAllocatedMetric
   { description :: Maybe Text
-  , enabled :: Bool
   , name :: Text
   , aggregate :: Maybe AggregationStrategy
+  , export :: Maybe ExportStrategy
   }
   deriving (Generic, Lift)
+
+instance HasField "enabled" HeapAllocatedMetric Bool where
+  getField :: HeapAllocatedMetric -> Bool
+  getField = isEnabled . (.export)
 
 instance FromJSON HeapAllocatedMetric where
   parseJSON :: Value -> Parser HeapAllocatedMetric
@@ -154,11 +249,15 @@ The configuration options for `GHC.Eventlog.Live.Machine.Analysis.Heap.processHe
 -}
 data HeapSizeMetric = HeapSizeMetric
   { description :: Maybe Text
-  , enabled :: Bool
   , name :: Text
   , aggregate :: Maybe AggregationStrategy
+  , export :: Maybe ExportStrategy
   }
   deriving (Generic, Lift)
+
+instance HasField "enabled" HeapSizeMetric Bool where
+  getField :: HeapSizeMetric -> Bool
+  getField = isEnabled . (.export)
 
 instance FromJSON HeapSizeMetric where
   parseJSON :: Value -> Parser HeapSizeMetric
@@ -175,11 +274,15 @@ The configuration options for `GHC.Eventlog.Live.Machine.Analysis.Heap.processBl
 -}
 data BlocksSizeMetric = BlocksSizeMetric
   { description :: Maybe Text
-  , enabled :: Bool
   , name :: Text
   , aggregate :: Maybe AggregationStrategy
+  , export :: Maybe ExportStrategy
   }
   deriving (Generic, Lift)
+
+instance HasField "enabled" BlocksSizeMetric Bool where
+  getField :: BlocksSizeMetric -> Bool
+  getField = isEnabled . (.export)
 
 instance FromJSON BlocksSizeMetric where
   parseJSON :: Value -> Parser BlocksSizeMetric
@@ -196,11 +299,15 @@ The configuration options for `GHC.Eventlog.Live.Machine.Analysis.Heap.processHe
 -}
 data HeapLiveMetric = HeapLiveMetric
   { description :: Maybe Text
-  , enabled :: Bool
   , name :: Text
   , aggregate :: Maybe AggregationStrategy
+  , export :: Maybe ExportStrategy
   }
   deriving (Generic, Lift)
+
+instance HasField "enabled" HeapLiveMetric Bool where
+  getField :: HeapLiveMetric -> Bool
+  getField = isEnabled . (.export)
 
 instance FromJSON HeapLiveMetric where
   parseJSON :: Value -> Parser HeapLiveMetric
@@ -217,11 +324,15 @@ The configuration options for the @memCurrent@ field `GHC.Eventlog.Live.Machine.
 -}
 data MemCurrentMetric = MemCurrentMetric
   { description :: Maybe Text
-  , enabled :: Bool
   , name :: Text
   , aggregate :: Maybe AggregationStrategy
+  , export :: Maybe ExportStrategy
   }
   deriving (Generic, Lift)
+
+instance HasField "enabled" MemCurrentMetric Bool where
+  getField :: MemCurrentMetric -> Bool
+  getField = isEnabled . (.export)
 
 instance FromJSON MemCurrentMetric where
   parseJSON :: Value -> Parser MemCurrentMetric
@@ -238,11 +349,15 @@ The configuration options for the @memNeeded@ field `GHC.Eventlog.Live.Machine.A
 -}
 data MemNeededMetric = MemNeededMetric
   { description :: Maybe Text
-  , enabled :: Bool
   , name :: Text
   , aggregate :: Maybe AggregationStrategy
+  , export :: Maybe ExportStrategy
   }
   deriving (Generic, Lift)
+
+instance HasField "enabled" MemNeededMetric Bool where
+  getField :: MemNeededMetric -> Bool
+  getField = isEnabled . (.export)
 
 instance FromJSON MemNeededMetric where
   parseJSON :: Value -> Parser MemNeededMetric
@@ -259,11 +374,15 @@ The configuration options for the @memReturned@ field `GHC.Eventlog.Live.Machine
 -}
 data MemReturnedMetric = MemReturnedMetric
   { description :: Maybe Text
-  , enabled :: Bool
   , name :: Text
   , aggregate :: Maybe AggregationStrategy
+  , export :: Maybe ExportStrategy
   }
   deriving (Generic, Lift)
+
+instance HasField "enabled" MemReturnedMetric Bool where
+  getField :: MemReturnedMetric -> Bool
+  getField = isEnabled . (.export)
 
 instance FromJSON MemReturnedMetric where
   parseJSON :: Value -> Parser MemReturnedMetric
@@ -280,11 +399,15 @@ The configuration options for `GHC.Eventlog.Live.Machine.Analysis.Heap.processHe
 -}
 data HeapProfSampleMetric = HeapProfSampleMetric
   { description :: Maybe Text
-  , enabled :: Bool
   , name :: Text
   , aggregate :: Maybe AggregationStrategy
+  , export :: Maybe ExportStrategy
   }
   deriving (Generic, Lift)
+
+instance HasField "enabled" HeapProfSampleMetric Bool where
+  getField :: HeapProfSampleMetric -> Bool
+  getField = isEnabled . (.export)
 
 instance FromJSON HeapProfSampleMetric where
   parseJSON :: Value -> Parser HeapProfSampleMetric
@@ -301,11 +424,15 @@ The configuration options for `GHC.Eventlog.Live.Machine.Analysis.Capability.pro
 -}
 data CapabilityUsageMetric = CapabilityUsageMetric
   { description :: Maybe Text
-  , enabled :: Bool
   , name :: Text
   , aggregate :: Maybe AggregationStrategy
+  , export :: Maybe ExportStrategy
   }
   deriving (Generic, Lift)
+
+instance HasField "enabled" CapabilityUsageMetric Bool where
+  getField :: CapabilityUsageMetric -> Bool
+  getField = isEnabled . (.export)
 
 instance FromJSON CapabilityUsageMetric where
   parseJSON :: Value -> Parser CapabilityUsageMetric
@@ -322,10 +449,14 @@ The configuration options for `GHC.Eventlog.Live.Machine.Analysis.Capability.pro
 -}
 data CapabilityUsageSpan = CapabilityUsageSpan
   { description :: Maybe Text
-  , enabled :: Bool
   , name :: Text
+  , export :: Maybe ExportStrategy
   }
   deriving (Generic, Lift)
+
+instance HasField "enabled" CapabilityUsageSpan Bool where
+  getField :: CapabilityUsageSpan -> Bool
+  getField = isEnabled . (.export)
 
 instance FromJSON CapabilityUsageSpan where
   parseJSON :: Value -> Parser CapabilityUsageSpan
@@ -342,10 +473,14 @@ The configuration options for `GHC.Eventlog.Live.Machine.Analysis.Thread.process
 -}
 data ThreadStateSpan = ThreadStateSpan
   { description :: Maybe Text
-  , enabled :: Bool
   , name :: Text
+  , export :: Maybe ExportStrategy
   }
   deriving (Generic, Lift)
+
+instance HasField "enabled" ThreadStateSpan Bool where
+  getField :: ThreadStateSpan -> Bool
+  getField = isEnabled . (.export)
 
 instance FromJSON ThreadStateSpan where
   parseJSON :: Value -> Parser ThreadStateSpan
@@ -376,17 +511,3 @@ encodingOptions =
     , tagSingleConstructors = False
     , unwrapUnaryRecords = False
     }
-
-{- |
-Internal helper.
-The encoding options that should be used by the `FromJSON` and `ToJSON` instances for `AggregationStrategy`.
--}
-aggregationStrategyEncodingOptions :: Options
-aggregationStrategyEncodingOptions =
-  encodingOptions
-    { constructorTagModifier = camelTo2 '_' . stripAggregationStrategy
-    , tagSingleConstructors = True
-    }
- where
-  stripAggregationStrategy :: String -> String
-  stripAggregationStrategy = drop (length ("AggregationStrategy" :: String))